backend/requirements.txt
requests
textblob
vaderSentiment
python-dotenv
nltk
sumy
newsapi-python
